Move loadNodes to its own file
