+++ packages/protocolbeat/src/store/actions/loadNodes.ts
@@ -1,9 +1,7 @@
 import { merge } from '../../api/merge'
-import { OklchColor } from '../../utils/color'
 import type { Node, State } from '../State'
 import { NODE_SPACING, NODE_WIDTH } from '../utils/constants'
 import { recallNodeState } from '../utils/localStore'
-import { type NodeLocations } from '../utils/storageParsing'
 import { updateNodePositions } from '../utils/updateNodePositions'
 
 export function loadNodes(
@@ -45,48 +43,6 @@ export function loadNodes(
     ...state,
     projectId,
     nodes: updatedNodes.concat(addedNodes),
-  })
-}
-
-export function hideSelected(state: State): Partial<State> {
-  return {
-    hidden: [...new Set([...state.hidden, ...state.selected])],
-    selected: [],
-  }
-}
-
-export function showHidden(): Partial<State> {
-  return { hidden: [] }
-}
-
-export function clear(): Partial<State> {
-  return {
-    projectId: '',
-    nodes: [],
-    hidden: [],
-    selected: [],
-  }
-}
-
-export function colorSelected(state: State, color: OklchColor): Partial<State> {
-  const nodes = state.nodes.map((node) =>
-    state.selected.includes(node.id) ? { ...node, color } : node,
-  )
-  return { nodes }
-}
-
-export function layout(state: State, locations: NodeLocations): Partial<State> {
-  const movedNodes = state.nodes.map((n) => ({
-    ...n,
-    box: {
-      ...n.box,
-      ...locations[n.id],
-    },
-  }))
-
-  return updateNodePositions({
-    ...state,
-    nodes: movedNodes,
   })
 }
